implemented header component to handle session status, login and logout
